Extract back navigation logic into hook
